% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brasilio.R
\name{brasilio}
\alias{brasilio}
\title{brasil.io covid data}
\usage{
brasilio(silent = !interactive())
}
\arguments{
\item{silent}{\code{logical} indicating if messages should be hidden.}
}
\value{
A \code{tibble} object.
}
\description{
Extracts updated data from the brasil.io page
by download the file in \code{.csv} format available and loading it into a comprehensive
\code{tibble} object.
}
