% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_ministerio_saude.R
\name{brMinisterioSaude}
\alias{brMinisterioSaude}
\title{Brazilian Health Ministry COVID data}
\usage{
brMinisterioSaude(silent = FALSE)
}
\arguments{
\item{silent}{\code{logical} indicating if messages should be hidden.}
}
\value{
A \code{tibble} object.
}
\description{
Extracts updated data from the page of Brazilian Health Ministry
(https://covid.saude.gov.br/)
by download the file in \code{.csv} format available and loading it into a comprehensive
\code{tibble} object.
}
