% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu.R
\name{CSSEGISandData}
\alias{CSSEGISandData}
\title{Data from the \code{CSSEGISandData/COVID-19} repository}
\usage{
CSSEGISandData(by_country = TRUE, silent = !interactive())
}
\arguments{
\item{silent}{\code{logical} indicating if messages should be hidden.}

\item{countries_only}{If \code{TRUE} the values are summed over the \code{Province.State} variable
for each \code{data} and \code{Country.Region}. In this case, \code{Lat} and \code{Long} are reported as the
mean value of all rows for that \code{Country.Region} value.}
}
\value{
A \code{tibble} object.
}
\description{
Extracts updated data from the repository by Johns Hopkins CSSE
(https://github.com/CSSEGISandData/COVID-19)
by download the file in \code{.csv} format available and loading it into a comprehensive
\code{tibble} object.
}
