% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoGripe.R
\name{infoGripe}
\alias{infoGripe}
\title{Data from http://info.gripe.fiocruz.br/}
\usage{
infoGripe()
}
\value{
A \code{list} with elements \code{casos} e \code{obitos}, containing \code{tibbles}
with the distribution of cases and deaths, respectively, by week and
age.
}
\description{
Extracts the data from http://info.gripe.fiocruz.br/ displayed
when the databases \code{SRAGCOVID} and \code{OBITOCOVID} are selected.
}
