% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registro_civil.R
\name{registro_civil}
\alias{registro_civil}
\title{Dados de https://transparencia.registrocivil.org.br/}
\usage{
registro_civil(state = "all", city_id = "all")
}
\arguments{
\item{state}{Estado no padrão \code{"SP"} ou \code{"all"} para o Brasil inteiro.}

\item{city_id}{ID de cidade. Uma tabela do ID por cidade pode ser encontrada
no objeto \code{registro_civil_cidades}.}
}
\value{
Um \code{tibble} com as mortes por gênero, faixa etária e causa nos anos de
2019 e 2020, do dia 16 de março até a data atual.
}
\description{
Dados de https://transparencia.registrocivil.org.br/
}
