% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{registro_civil_cidades}
\alias{registro_civil_cidades}
\title{Tabela de city_id para registro_civil()}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5610 rows and 3 columns.
}
\usage{
registro_civil_cidades
}
\description{
Tabela com os IDs de cidades para usar
com a função \code{registro_civil()}.
}
\keyword{datasets}
