% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brasilio.R
\name{brasilio}
\alias{brasilio}
\title{brasil.io covid data}
\usage{
brasilio(silent = !interactive(), cache = FALSE, invalidate_after = 12)
}
\arguments{
\item{silent}{\code{logical} indicating if messages should be hidden.}

\item{cache}{\code{logical} indicating if cached results should be used (if they exist).}

\item{invalidate_after}{The time (in hours) to invalidate the cache.}
}
\value{
A \code{tibble} object.
}
\description{
Extracts updated data from the brasil.io page
by download the file in \code{.csv} format available and loading it into a comprehensive
\code{tibble} object.
}
