% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ministeriosaude.R
\name{brMinisterioSaude}
\alias{brMinisterioSaude}
\title{Brazilian Health Ministry COVID data}
\usage{
brMinisterioSaude(silent = !interactive())
}
\arguments{
\item{silent}{\code{logical} indicating if messages should be hidden.}
}
\value{
A \code{tibble} object.
}
\description{
Extracts updated data from Brazilian Health Ministry website
(https://covid.saude.gov.br/)
by downloading the file in \code{.csv}/\code{.xlsx} format available and loading it into a comprehensive
\code{tibble} object.
}
