% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu.R
\name{CSSEGISandData}
\alias{CSSEGISandData}
\title{Data from the \code{CSSEGISandData/COVID-19} repository}
\usage{
CSSEGISandData(silent = FALSE)
}
\arguments{
\item{silent}{\code{logical} indicating if messages should be hidden.}
}
\value{
A \code{tibble} object.
}
\description{
Extracts updated data from the repository by Johns Hopkins CSSE
(https://github.com/CSSEGISandData/COVID-19)
by download the file in \code{.csv} format available and loading it into a comprehensive
\code{tibble} object.
}
